+++ src/main/java/data_access/InMemoryUserDataAccessObject.java
@@ -1,14 +1,19 @@
 package data_access;
 
 import entity.User;
-import session.SessionManager;
 import use_case.execute_buy.ExecuteBuyDataAccessInterface;
+import utility.SessionManager;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * A class that implements the ExecuteBuyDataAccessInterface interface
  * This class is used to get the user with the given credential
  */
 public class InMemoryUserDataAccessObject implements ExecuteBuyDataAccessInterface {
+    private Map<String, User> users;
+
     @Override
     public User getUserWithCredential(String credential) throws ValidationException {
         SessionManager sessionManager = SessionManager.Instance();
@@ -16,8 +21,21 @@ public class InMemoryUserDataAccessObject implements ExecuteBuyDataAccessInterfa
         return getUserWithUsername(username);
     }
 
-    // TODO: implement this method, it should read from an in-memory database
     private User getUserWithUsername(String username) {
-        return new User(username, "0");
+        Map<String, User> users = new HashMap<>();
+
+        User user1 = new User("user1", "password");
+        user1.addBalance(10000.00);
+        users.put(username, user1);
+
+        User user2 = new User("user2", "password");
+        user2.addBalance(20000.00);
+        users.put(username, user2);
+
+        User user3 = new User("user3", "password");
+        user3.addBalance(30000.00);
+        users.put(username, user3);
+
+        return users.get(username);
     }
 }