+++ docs/CHANGELOG.md
@@ -34,5 +34,6 @@ Implemented `ClientSessionManager` for client-side session management.
 - Implemented buy use case interactor and set up unit tests
 - Removed outdated classes and methods that are incompatible with current codebase version
 - Updated maven configuration with unit test and api request dependencies
+- Implemented the frontend framework
 
 ### Bug Fixes